Move component generation to a command.
